requirements.txt
Flask
pdfplumber
pytesseract
python-docx
openai
werkzeug
